// InkeyCountryStateCityDropdownControl/generated/ManifestTypes.d.ts
/*
*This is auto generated from the ControlManifest.Input.xml file
*/

// Define IInputs and IOutputs Type. They should match with ControlManifest.
export interface IInputs {
    value: ComponentFramework.PropertyTypes.StringProperty;
    mode: ComponentFramework.PropertyTypes.StringProperty;
    country: ComponentFramework.PropertyTypes.StringProperty;
    state: ComponentFramework.PropertyTypes.StringProperty;
}
export interface IOutputs {
    value?: string;
}